+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchEvent.kt
@@ -1,0 +1,5 @@
+package org.simple.clinic.home.overdue.search
+
+sealed class OverdueSearchEvent
+
+data class OverdueSearchHistoryLoaded(val searchHistory: Set<String>) : OverdueSearchEvent()